ref: extract patch preview
